#1890: extract coordinator/tunnel_supervision.rs (GRE+WG thread lifecycle)
coordinator/mod.rs crossed the ~2,000 prod-LOC modularity threshold
(2,544 lines after #1887). First of three pure code-motion extractions
per the corrected split map on the issue: move the GRE local-origin and
WireGuard control-thread supervision block (the #1887-reworked
three-pass lifecycle — finished sweep, stale prune, spawn with backoff,
plus the periodic liveness sweeps and defer-branch snapshot prunes)
into coordinator/tunnel_supervision.rs, together with the
WG_SPAWN_BACKOFF_NS backoff constant that only this block uses.

Pure motion: method bodies, doc comments, and the const are
byte-identical to the mod.rs originals. The only deltas are the two
visibility widenings the move forces — reconcile_local_tunnel_sources
and spawn_wg_control_threads go from private-in-coordinator to
pub(super) (same coordinator-subtree audience: their callers
reconcile/bringup.rs and the snapshot-refresh leg become sibling files
of the new module). All other helpers stay private to the new file;
the pub(crate) surface is unchanged. Entry maps (tunnel_sources,
wg_control_threads) remain fields on Coordinator and stop_inner keeps
its inline teardown in mod.rs.

Validation: cargo build --release clean; warning set byte-identical to
the origin/master baseline (diff of sorted warning lines). Full test
gates run at the end of the series.
